Add feature flags library
We want to be able to feature flag new code to be able to better control
how we release new things.

The govuk_feature_flags library is in use in other projects (Find +
Refer) and has proven to be a simple way of adding this functionality.

This change includes an update to the layouts in order to be able to
display a separate navigation for the support section versus the public
view.