Add handling of invalid tokens to QualificationsApi::Client#teacher
If a token is badly formed or expired, the Qualifications API returns a
HTTP 401. Error messages in the response headers distinguish these
cases, but for our purposes we can treat them the same way for now.
Specifically, by raising a QualificationsApi::InvalidTokenError.

In the client we:

- Look for a 200 response, upon which we return the parsed teacher data
- Look for a 401 response, upon which we raise the InvalidTokenError

We may want to implement something that handles all other cases, but
that can follow in a later commit.