Add the UI for qualifications
The prototype has a [design for the qualifications screen](https://qualifications-prototype.herokuapp.com/landing-pages/v2/qualifications).

We want to recreate this in the app.

A lot of the data being displayed isn't present as it will be fetched
from the Qualifications API.

In order to allow us to build the views, I opted to stub out the data
with values from the prototype. The aim would be to replace these with
the data from the API once that connection is setup.

I upgraded the `govuk-components` dependency to the `4.0.0a2` version so
that we have access to the new summary card component.

This requires a few changes elsewhere in the views due to breaking
changes there.

Also, I introduced the `name_of_person` library for the convenience
helpers it provides around displaying names.