Add basic starting point for auth journey
The design proposed in the prototype has the starting point for the
Qualifications auth journey be either the GOVUK start page or an
interstitial page on Identity. Neither of these are currently built.

For now, have the root path of the service be a rudimentary sign in
page. This commit also includes before_action filters to ensure users
are correctly redirected to either the sign in page or the
qualifications screen depending on their auth state.