Add a common interface for qualifications
The DQT API returns a variety of qualification types and uses separate
attributes on the response to represent them.

The UI we are building treats displays all qualifications in the same
format.

I opted to have the API client transform the incoming data into a
consistent interface that makes the UI simpler and easier to work with.