Add route constraint for AYTQ service
We're planning to build a Check teacher records service. We've decided
to implement it as part of this repo.

Following a pattern established by DFE-Digital/publish-teacher-training,
we can use route constraints to ensure routes for a given service are
only accessible via requests to the appropriate hostname. This allows
the two services to live at different URLs and use the same codebase
without any conflicts.

In this commit:

- Introduce the route constraint for AYTQ
- Move AYTQ routes into their own file, and apply the constraint
- Assume for now that support / staff routes will be shared between the
  two services
- Add HOSTING_DOMAIN to .env.test so that system specs pass