Fetch Induction data from the API
The DQT API now supplies data on the teacher's induction qualification.

This data differs from the other qualifications and due to the way the
GOVUK view components are built, we need to create a custom version of
the summary list component. Allowing us to accommodate the induction
history entries.

The exceptions added to Rubocop in this change are due to conflicting
rules in prettier-rb.

These two linters take the opposite approach when it comes to multi-line
blocks. I disabled the Rubocop ones in favour of prettier because the
syntax-tree autocorrect changes this every time you save the offending
file causing constant churn of these lines.