Display any mandatory qualifications
The DQT API is now providing details of the mandatory qualifications.

We can display these and the one difference is that they don't have an
associated certificate.

I opted to simply extend the existing QualificationSummaryComponent to
handle this edge case. We don't expect any more additions to the API
that we would need to handle.

If this changes in the future, we would consider making separate view
components for different qualifications.