Setup DB encryption for PII fields
We store some PII data in the `users` table and want to add encryption
as an extra layer of protection in case someone gained unauthorized
access to the database.

Rails has good support for encrypting fields that we can leverage here.

This change introduces the encryption for all non-production
environments.

A future PR will add support for production.

Also, this change doesn't actually encrypt any existing data. That will
happen in a separate change. To allow us to migrate existing data, a
config value is set to support unencrypted data.

We will remove this configuration once all data is migrated.

Fortunately, there isn't any "real" data to migrate, so the risk here is
very small.