Add data migrations
There are times when we want to migrate data across all environments but
don't want to place these operations in a schema migration.

The `data_migrate` gem provides this functionality.

To use it, run the generator `rails g data_migration name_of_migration`.

This creates a file in `db/data`.

Then run `rails data:migrate` to execute the migration.