Add console1984 & audits1984
We have encrypted user PII now and want to protect access to it via the
Rails console.

There are 2 libraries that can help us with this. `console1984` provides
the ability to decrypt and encrypt the data.

`audits1984` provides auditing of these decryption actions along with an
interface to review the logs.

Something to note is that we aren't able to connect the auditing to our
authentication system.

The console asks for details about the person making the decrypt request
along with details of consent to view it. However, this is entirely
reliant on the user providing truthful details here.

This doesn't protect against bad actors.