Add support for LogStash
We want to make use of LogStash to access logs in production.

In order to do this, we need to ensure the logs are in a suitable
format.

Firstly, we need to convert the verbose logging default Rails format
into a more compact JSON format.

`rails_semantic_logger` helps with that.

There's a wrinkle in this process in that the way exceptions are handled
by semantic logger isn't supported by LogStash. Adding a custom format
for this is required.

Secondly, we conditionally send the newly formatted logs to LogStash
using the presence of the environment variables as the gate.

Assumptions:

* LogStash will be configured on the project shortly by the DevOps team
* Access to the logs will be managed via Azure
* The condensed format doesn't interfere with how devs currently use the
  logs