Update Identity auth to work without Devise
We're going to add DfE Sign In to this codebase for use with the Check
teacher records service. Devise has issues supporting more than one
omniauth-based authentication flow.

To address this, we can remove our dependency on Devise for the Identity
auth flow by not using its helpers or inheriting from its controllers.
We can then eventually adopt a similar approach for DSI.