Fix identity account link
The account link doesn't work due to using a form instead of a link.

This resulted in the URL params not being passed correctly.

Switching to a link that is styled as per the designs fixes this.

While making this change, I also added the sign out URI value to enable
the sign out flow.