Pass a session ID to Identity
We want to be able to track a session across apps in order to better
deliver a good authentication experience.

The identity provider will add a provided `session_id` to every tracking
event it sends to BigQuery.

This will allow us to better connect sessions across different BigQuery
tables.

The `dfe-analytics` library doesn't have support for custom data in the
web request event type.

Rather than extend that library, as it appears this is not something
required across other apps yet. I opted to monkey-patch the method that
triggers the web request event to add custom data.

We're using the Rails session ID as the identifier to be used across
apps as it is the most unique thing we can use across services.