Add session expiry for DSI authentication
When authenticating in the check records service using the DSI method,
there is no token expiry.

A long-lived token like this has an increased security risk over a
short-lived one.

To resolve this, we are proposing adding an arbitrary time out on the
session, starting with a 2 hour limit.

There is nothing magic about the choice of 2 hours, so if it turns out
that you think there is a better window, then feel free to change.

The implementation sets a hard-limit of 2 hours from a successful
response from the identity auth provider. This means that if the user
has a single session longer than 2 hours, they will be required to sign
in again to continue accessing the service.