Encrypt sensitive DSI User fields
Whenever we store sensitive data on an individual, we want to encrypt
it.

This change adds encryption to the DSI User record for PII. Currently,
the only affected fields are email, first name and last name.

The encrypted version of the fields are longer and don't reliably fit
within the default 255 character limit, so we increase the length of the
column to accommodate this.