Update Quals API client to support /v3/teachers/{trn} endpoint
The Check records service will use this endpoint, in conjunction with a
fixed bearer token, to retrieve teacher details via their TRN.

Add support for the endpoint to the existing Quals API client.

As part of this change, use metadata in our specs to make it more
obvious that we're using the Quals API fake.