Add Sentry for error tracking
We want a way to manage app errors.

Sentry is the standard service we use across multiple projects for this.

The `SENTRY_DSN` value that the library requires needs to be set via an
ENV variable.

This change assumes that this will be provided from the Azure keyvault
and no further setup needs to happen here.