Change CHECK_RECORDS_DOMAIN in development
This needs to be different from the AYTQ domain in order for root path
helpers to work properly, which in turn is needed for the DSI auth flow
for Check to make sense.

Some detail around routing and the auth flow will be documented in a
later commit. For now, include this change to encourage the use of
different hostnames when developing locally (ie - use
http://check.localhost:3000/ when accessing Check in development).