Fix QTS certificate downloads
Some code was introduced which handles certificate download URLs in a
generalised way for all certificates, including code dealing with the
certificate id.

QTS certificate URLs don't contain an id so this was resulting in a
slightly incorrect URL being used when retrieving data from the API.

Add a conditional check in Qualifications#id specifically for QTS
certificates, so that we use the correct API endpoint.