Separate assets between services
The services currently share assets but we'd rather they didn't, so as
to avoid users fetching code that never runs.

I opted to do a simple duplication of the stylesheets and JS for now.
The assumption is that they will diverge over time as specific styles
and functionality are added.

Note that the `application` layout still remains for use by the admin.