Hide ITT row if no data
The current implementation of the qualifications view assumes that there
will always be an ITT record for a teacher.

There are scenarios where this isn't true. This results in the component
displaying with empty values, which is confusing for the user.

Instead, we want to conditionally render the ITT row only if it is
present in the returned data.