Fix EYTS downloads
EYTS downloads will be blank with the current implementation. This is
the same issue that affected QTS certificates recently, namely that
their URLs don't contain ids, so were being incorrectly handled by code
that infers the API endpoint to retrieve certificate data from.

This commit applies a conditional check for EYTS in the Qualification model
before deciding whether or not to return an id.