Implement RP-initiated logout for OAuth provider
The Identity service implements the OpenIDConnect spec, which allows for
an RP-Initiated logout.

We aren't using the OpenIDConnect OmniAuth strategy currently, so don't
have access to the logout feature.

I considered re-implementing the strategy to use OpenIDConnect but
decided that it seemed overkill for what we need here, which is to
support a redirect on sign out.

Instead, I implemented the `other_phase` method that all OmniAuth
strategies call for non-matching requests.

We can match on the logout URL and trigger the correct redirect.