Add generic OAuth failure handling
We need some basic error handling in the event of a failure during the
OAuth flow, in both Identity and DSI auth strategies.

This commit adds a failure controller which simply redirects the user to
the correct root path, depending on which auth strategy they're using.

The primary error scenario this handles is when users navigate back in
their browser to a stale page on the auth provider site. These cases
don't provide particularly helpful error messages, so I've elected not
to display an error message to the user.