Add support for qualifications with a missing date
There are scenarios where a qualification has no associated date.

Currently, this causes the app to throw an exception as we sort
qualifications by date.

I landed on using a default value for the date in the sort function.
This allows us to position the qualifications without a date at the end
of the list, so it makes sense visually.

We don't want this date to be visible though as it is used purely for
the sort order, so I didn't set the default on the qualification itself.