Add issuer value to DSI config
This will tell the omniauth gem where to find openid configuration data
during the discovery phase. It will look for it at:

https://issuer-url/.well-known/openid-configuration

Without this, the default behaviour is to use a Webfinger endpoint, ie:

https://issuer-url/.well-known/webfinger (with some query params
describing the resource)

The OIDC implementation in DSI no longer supports the Webfinger
protocol, hence the switch.