Remove default values for OIDC configuration
In a [discussion on a different PR](https://github.com/DFE-Digital/access-your-teaching-qualifications/pull/234#discussion_r1260947743),
there was a suggestion
to remove the default values from the OIDC configuration to trigger
misconfiguration messages sooner, eg. on app startup.

The idea is that this would make development better as any
misconfiguration error would be clearer.