Set worker count to zero
Temporarily deploy the service without workers

We only use the workers to send analytics data to BigQuery. BQ isn't
fully configured yet. This will allow the jobs to queue up but not be
sent until we re-enable the workers. We may need to rethink this if it
takes much longer to set BigQuery up.