Check organisation on sign in
When a DSI user signs in to the service, we want to ensure that they are
part of a known organisation before letting them access the service.

The list of known organisations is a CSV provided by the Identity team.

At this stage we don't know how frequently this list will change.

The company registration number is the proposed preferred field to match
on.

Testing this requires being added to the test organisation in
development or pre-prod environments of Identity. You can request this
from the DSI team.