Use the provided certificate URL
Currently we construct the URL for a certificate by using the
certificate type.

There are reports from users of 404 pages and the exceptions point to a
timeout when fetching a certificate.

The working theory is that we're constructing a URL that doesn't exist
and so the API call hits the timeout.

There's no reason to construct the certificate URL in this fashion as
the API returns the path if the certificate is present.

We can rely on this provided value instead.

If no certificate is returned we want to display a notice to the user
that the certificate could not be found.