Make .env.development domain values more consistent
Use full URLs for both HOSTING_DOMAIN and CHECK_RECORDS_DOMAIN,
including port number. This makes the two values structurally consistent
with each other and with the values we use on the deployed environments.

If a different port number is needed for local development, either env
var can be overridden in .env.development.local.