Allow qualifications without an awarded at value
We're seeing a handful of errors related to rendering qualifications
with a missing awarded at value.

We don't control the values returned by the API but it is possible that
there are edge cases where an awarded at date is missing but we still
want to display the qualification.

I opted for a defensive approach, where we simply guard against the nil
but a future iteration of this may include displaying a default when the
value is missing.