Display restrictions on the search detail
When viewing the details of a returned search result, we want to also
see the restrictions badge.

The designs have extra information that isn't currently available via
the API. I've left those out for this round of changes.

Once the API provides the extra information, we can revisit this and
display them too.