Add a basic support interface for adding organisations
We have an `Organisation` model that is used to control access to the
system.

Currently, there is no way to add an organisation record.

This change introduces a very simple interface into the support section
to allow us to create a record. The intention here is that it will allow
us to create a record in development for testing.

The approach wouldn't be pleasant if you had to create more than a
handful of records, which we will need to do in production. A different
approach will be required for that.