Add basic auth bypass to Check service review apps
Repurpose the existing dev bypass and make it available in review apps
as well.

We can iterate on this and eventually introduce different kinds of user
sign in. Currently we don't have any kind of permissions model to
emulate, so here we just click through and sign in as a generic user.
This gives full access to Check functionality.