Display sanction details for a specific teacher
When a teacher has sanctions in the API response, we want to display
them as per the designs.

There is on-going work to update the wording used currently to describe each
sanction. See https://educationgovuk.sharepoint.com/:x:/s/TeacherServices/EZsyllhdkb9Giji9pHvGG1UBKseSEplO-KUUH-vseaRM7w
for details.

We don't have a full description or date attached to any of the
sanctions at the moment, so this is where the implementation diverges
from the designs.

Assumptions:
- the sanctions list won't change much so storing them as key-values in
  code will be the most efficient way of representing them
- we won't be getting any more details from the API about the sanctions
- we can ship a version of the sanctions without the descriptions and
  revisit this once they have been written