Fix Staff email invite
There is an issue when inviting a staff user where the check for the
AnonymousSupportUser class doesn't work as expected.

Rather than check on class, we can use duck-typing and check for the
presence of an expected method.

This means the code is less coupled to a specific class and is more
explicit about the expectations for the code branch.