Configure sign out of DSI from check records service.
Adds a OmniAuth strategy which includes the id_token_hint param in the
end_session_redirect_uri. Necessary for the node-oidc-provider library
to end the session and redirect successfully.

The id_token_hint value is the auth.credentials.id_token returned on
successful sign in. This value is stored in the user session and passed
as a param in the DfE signout path.

Configures omniauth to use this strategy.

Adding the redirect route /check-records/auth/dfe/sign-out is also
necessary to make the redirect work as the underlying omniauth_openid_connect
gem pattern matches on the request_path as part of the redirect
validation.