Fix the change password flow
There is an error in the change password flow where the reset token
doesn't get correctly passed to the model.

It seems that the controller expects a non-nested location in the params
for the reset token. Our version nests it inside a staff namespace.

While making this change, I added better support for displaying errors
on the change password page.