Redirect the user through the DSI logout flow on session expiry
If a user attempts to go back to the DSI org selection step in the auth flow, a DSI sessionexpired error failure will occur.
There's no reasonable way to navigate back to this step so instead, redirect them through the DSI signout flow which ultimately takes them back to the sign in page.