Format date errors according to the style guide
The GOVUK design system describes specific error styling for date fields
that the current helpers don't support.

Due to the way we build the errors and Rails conventions, there is no
support for highlighting errors with specific parts of a date input.

Adding support for this requires a few changes in the form builder.

I considered adding these changes upstream in the form builder gem but
opted instead to monkey-patch them initially so that we can explore them
in a real environment first.

Once we have seen these changes work, I will propose the changes to the
form library and remove the monkey-patch if they are accepted.

Mostly, the changes are around how we determine if an error message is
present for one of the date part input fields.