Sort qualifications by career timeline
Sorts qualifications by their type according to the typical career
progression of a teacher.
ie.
NPQ or MQ | Induction | QTS | ITT (for QTS) | EYTS | ITT (for EYTS) | Mandatory | HE

This produces a consistent order for the types of qualification starting
with the most recent.
In the case of a specific type of qualification having multiple results,
the qualifications are sorted by reverse chronological awarded at date.