Remove restrictions tag
We have decided to remove the restrictions tag from both the search
results page and the detail page for a teacher.

The reasoning is that it doesn't add any useful information at best and is
potentially confusing at worst.

We also want to remove the date of the sanction.

See [Trello](https://trello.com/c/zlTNQhY4/133-update-restrictions-display) for the details.