Update rendering of NPQ names in Check
Currently the Qualifications API returns the NPQ name in the following
form:

"National Professional Qualification (NPQ) for [name of NPQ]"

We want to remove the acronym expansion at the start so we can use the
actual name of the qualification in a field heading.

Add a gsub to remove the prefix if it is present.