Add description to sanctions
We have expanded the sanctions list to also display a description
alongside the title.

This data isn't provided by the API so we opted to define it the same
way we do with the titles.

The descriptions contain links and we want them to be clickable so
using markdown made sense here.

The library is already being used elsewhere in the app, so there is no
extra dependency required. Also, we control the markdown text itself, so
there is no security risk by rendering this using the `html_safe`
method.

The use of `CHOMP` on the heredoc is to ensure we avoid the trailing
newline character that a heredoc automatically adds.