Add support for previous last names
The Qualifications API can return a person's previous names if
requested.

Our designs now include this in the detail view of a person.

It is a simple matter of requesting the extra data in the API request
and then ensuring the resulting output matches the format of the design,
eg. each previous last name on a separate line.