Use the statusDescription field for induction
The status field for the induction record is an enum that we don't
properly use in Check.

Instead of duplicating the values, the API now provides the human
friendly version of the enum as a new field, statusDescription.

This allows us to remove some code related to displaying the old field.