Change null certificate URL behaviour in fake quals data
The fake API data module has a facility by which the certificate URLs
are set to nil if a TRN is explicitly provided to the `quals_data`
method. This doesn't appear to be used anywhere in the test suite - we
appear to simply set the nil value manually when the test scenario calls
for it.

Because it's a little unintuitive, and because we're about to add some
functionality to the API test data to deal with null values, we can
remove this option from the existing FakeQualificationsData module.