Add spec and test data for nullable fields in the Qualifications API
Some fields from the Qualifications API response are nullable. We were
handling these inconsistently which led to bugs in production AYTQ.

This commit adds test data simulating a response where most nullable
fields are empty. The system spec checks that the page renders without
error.

This is implemented as follows:

- Start with the full response generated by the
  FakeQualificationsData module.
- Apply a series of mutations which nil out any fields that are
  nullable according to the Swagger docs for the v3 API.
- We don't nil out entire qualification fields, eg the qts field. Data
  is returned, but only the minimal subset according to the API docs.
  This reflects the kind of error we've actually encountered in
  production. More detailed testing of empty qualifications can be
  handled in unit tests.
- Make the altered response data available via
  FakeQualificationsDataWithNulling.generate.

The FakeQualificationsApi has been updated to support an additional
token "nulled-quals-data", which returns the data described above in the
system spec.