Skip start page, redirect users to ID
Skips the start page by emulating the same request made by
the start button in the appropriate controller action.
Redirects users to the location returned by the OmniAuth request phase call to the identity service.
Checks that this redirect location is a valid identity service URL.
Renders the existing start page if Faraday encounters an error when making
the call.