Add additional detail around 403s from Qualifications API
We haven't been handling 403 errors from the API, and are starting to
see quite a few in production. It isn't currently clear what's causing
these errors.

This commit adds a local Sentry scope that extracts various pieces of
information about the Identity session and sends it to Sentry whenever
this error occurs. This should help us with debugging the issue.