Redirect root to a separate signin page
We are going to move the start page to GOV.UK and so don't want users to
see another full start page when they land on the service.

But... the existing one is still the first result on Google so we don't
want those users to get the cut down login page as they won't have seen
the content on GOV.UK.

There is an existing slimmed down sign-in page that doesn't seem to be
linked from anywhere so redirect to that when users land on root but
leave the /qualifications/start in place until Google drops it from its
index. We can redirect it then.