Return support_interface namespace for FeatureFlagsController
The DsiAuthenicatable controller concern needs to construct some dynamic paths for the various namespaces in the app.
To do this it uses the underlying module name of the including controller, but in the case of feature flags, the module name is FeatureFlags
because this module comes from the govuk_feature_flags gem.
Return :support_interface in this case.