Add CSRF error handling to AYTQ sign out controller
If the user clicks the back button after sign out and submits the
Confirm sign out form again, they'll encounter a CSRF error (because their
session has already been cleared).

Add error handling to the Qualifications::Users::SignOutController that
takes the user back to the start page in the event of an
InvalidAuthenticityToken error.