Hide name change link if request pending
The TRS API returns a field pendingNameChange set to true if the user
has already submitted a name change request. Look this up and:

- Render a Change link if the field is false
- Render an 'In Review' tag if the field is true

I've opted to reimplement the summary list containing name/DoB in pure
HTML rather than a helper method as it's not obvious how to embed a tag
inside the actions column using the govuk-components summary_list helper.