Encrypt User#one_login_verified_name
This is PII, so we should encrypt it.

The One Login feature hasn't been activated in production yet, so we can
adopt a fairly simple migration approach:

- Truncate this column in the database
- Specify that this is an encrypted field in the User model

If we don't truncate, attempts to read or update existing User records
in the dev/test environment will cause an error (because the field will
contain unencrypted data that Rails will try and decrypt). We could
configure ActiveRecord Encryption to accept both encrypted and
unencrypted data but this seems needlessly obtuse compared to the
approach above.