Add redirect if user submits repeat change request
When a user submits a name or date of birth change request, they
shouldn't be able to submit another until the pending one has been
processed.

We deal with this in the UI by removing the 'change' link on the account
page if the user has a change pending. They can still submit a repeat
request, however, by hitting the back button after submitting a form and
then resubmitting.

This commit adds a check for the pending state of both name and date of
birth changes, redirecting the user back to the account page if
appropriate.