Fix redirects from name and date of birth change controllers
We have before filters in these controllers that redirect the user back
to the account page if they attempt to resubmit a name or date of birth
change form. These filters were applying only to the create/update
actions, meaning that a user could still resubmit a request from the
confirmation page.

Remove the whitelist on these filters so that all controller actions are
inaccessible if a request is pending.

Amend the error message wording slightly while we're at it.