Add TRN search to CTR
Users in CTR search by last name and date of birth. This can sometimes
result in multiple search results. For data protection reasons, we want
to reduce cases of users seeing multiple results.

Address this initially by adding a TRN search page when multiple results
are returned by the API.

The user should be able to skip this page, functionality which will
appear in a later commit.