Correct trn search behaviour
Currently the TRN search in CTR does an API call for a teacher with the
submitted TRN value. This isn't the desired behaviour. We should be
doing a last name / date of birth search, and then filtering on this
result set.

Alter the logic of CheckRecords::SearchController#trn_result to reflect
this.