Restrict feedback to authenticated users
We are getting spam in the feedback form and want to reduce this.

The proposal is to restrict access to the feedback form using the
authentication system.

In this first iteration, we are redirecting to the start page of the
service and showing a flash message with instructions about needing to
sign in to give feedback.