Redirect to feedback form after sign in
We want to make the flow of signing in to provide feedback easier.

Currently, if you click the feedback link while signed out, you are
returned to the start page and prompted to sign in. If you then proceed
to sign in, you need to click the feedback link again after successfully
authenticating.

As a first step to making this flow have less steps for the user, I
opted to store the intention of providing feedback as a `return_to`
value in the session.

The user is still prompted to sign in but automatically gets redirected
to the feedback form if that's what they were attempting to do
previously.