Add a warning message for when there are no details
There are occasions when we don't have any details on a person even
though we have a record of them.

In these scenarios, to make it clear to users that there was a match, we
want to present the warning text that there are no details on record for
the matched person.