Add support for programme type as an Integer
We have a new programme type returned by the DQT API that is an Integer
rather than a String.

This breaks the code that partitions the qualifications as it expects a
String.

We currently have no use for the Integer value so I've opted for the
simplest change of converting it to a String.