Enable searching for multiple records
We want to provide a way for people to search for multiple records in a
single request.

This introduces the concept of using a CSV to define multiple search
queries. We also have access to a new API endpoint that accepts multiple
queries and returns matches in a single request.

The designs can be viewed in [Figma](https://www.figma.com/design/rBTUYusPtlj80dENohuoVJ/CTR---Work-in-progress?node-id=684-11598&m=dev).

The new UI provides guidance around how to format the CSV and then
some validation around the contents of the CSV prior to making the API
request.

We opted to validate before making the request as we can tie the error
messages to rows in the CSV more easily that way and also it is quicker
as we don't need to make an API call to validate missing or invalid
formatting of data.