Move the timestamp representation to a component
We display the timestamp of when an action was taken in a few places and
the implementation varies slightly in each spot.

I opted to move this to a ViewComponent to make it easier to test this
component in isolation.

Also, we can use the DATE_FORMATS setting to avoid hard-coding date
formats across the code and tests.