Explicitly set the timezone
When display the time an action was performed, it is consistently being
rendered in UTC.

Previous efforts have included setting the Time.zone in the application
configuration but this hasn't worked.

Given that we have a single place where we calculate the timestamp, we
can explicitly set the time zone to London to ensure it displays
correctly.