Upgrade data_migrate gem for Rails 7.2 compatibility
The version of data_migrate (9.4.2) that is currently installed
references a method on the database connection that is removed in Rails
7.2.

This prevents us from being able to deploy the app as booting fails on
the data migration step.

There is no rubygems version that supports Rails 7.2 yet, so I opted to
switch to referencing the repo on GitHub directly.

We can revisit after a new release is published. This is looking like it
will coincide with the release of v11.