Send a custom event on validation errors
When a user sees a validation error while trying to perform a bulk
search there is no corresponding event tracked in BigQuery. This makes
it hard to analyse how the service is used and how we could improve it.

Adding a custom event that contains the validation error message is a
first step in being able to bring some observability to this particular
experience.

The event simply contains the full validation error message. We will
launch the feature with this and can iterate on it as needed.