Handle API errors for bulk search
There are cases when the API throws an error that we can handle better
for the user.

If an error is returned the body of the response is empty, which leads
to an exception due to the expectation of there always being a body
value.

I opted to guard against the nil value to ensure we can always render a
response.

For the error handling, I opted to raise an exception with the error
code and message so we can get insight into the issues during testing.